+++ aeweb.js
@@ -5,7 +5,7 @@ import { hideBin } from 'yargs/helpers'
 import about from './commands/about.js'
 import generate_address from './commands/generate_address.js'
 import deploy from './commands/deploy.js'
-import AEWeb from './api.js'
+import AEWeb from './lib/api.js'
 
 const y = yargs(hideBin(process.argv))
