+++ commands/cli.js
